added file upload

+++ application/controllers/Contatos.php
@@ -15,8 +15,14 @@ class Contatos extends CI_Controller {
 	}
 	public function delete($id){
 		if(!empty($id) && is_numeric($id)){
-			$this->Contatos_model->delete(array('id'=>$id));
-			$this->session->set_flashdata('mensagem','Contato removido com sucesso');
+		    $contato = $this->Contatos_model->get_where(array('id'=>$id));
+		    if($contato) {
+		        unlink('uploads/'.$contato->arquivo);
+                $this->Contatos_model->delete(array('id' => $id));
+                $this->session->set_flashdata('mensagem', 'Contato removido com sucesso');
+            }else{
+                $this->session->set_flashdata('mensagem', 'Contato não existe.');
+            }
 		}else{
 			$this->session->set_flashdata('mensagem','Contato não encontrado');
 		}
@@ -31,8 +37,7 @@ class Contatos extends CI_Controller {
 		$this->load->view('contatos/form',$data);
 	}
 	public function create_action(){
-		$this->form_validation->set_rules('nome', 'Nome', 'required');
-		$this->form_validation->set_rules('email', 'Email', 'required');
+		$this->_validationRules();
 		if ($this->form_validation->run() == FALSE){
 			$this->cadastrar();
 		}else{
@@ -40,8 +45,18 @@ class Contatos extends CI_Controller {
 				'nome'=>$this->input->post('nome'),
 				'email'=>$this->input->post('email'),
 			);
-			if($this->Contatos_model->insert($insert)){
-				$this->session->set_flashdata('mensagem','Contato cadastrado com sucesso');
+			$contato = $this->Contatos_model->insert($insert);
+			if($contato){
+               $this->_configsUpload();
+                if (!$this->upload->do_upload('arquivo')){
+                    $this->session->set_flashdata('mensagem',$this->upload->display_errors());
+                    redirect(base_url('contatos'));
+                    exit();
+                }else{
+                  $upload = $this->upload->data();
+                  $this->Contatos_model->update($contato,array('arquivo'=>$upload['file_name']));
+                }
+          		$this->session->set_flashdata('mensagem','Contato cadastrado com sucesso');
 			}else{
 				$this->session->set_flashdata('mensagem','Falha ao cadastrar o contato');
 			}
@@ -56,6 +71,7 @@ class Contatos extends CI_Controller {
 			$data['action'] = base_url('contatos/update_action/'.$contato->id);
 			$data['nome'] = set_value('nome',$contato->nome);
 			$data['email'] = set_value('email',$contato->email);
+			$data['arquivo'] = $contato->arquivo;
 			$this->load->view('contatos/form',$data);
 		}else{
 			$this->session->set_flashdata('mensagem','Contato não encontrado.');
@@ -63,15 +79,28 @@ class Contatos extends CI_Controller {
 		}
 	}
 	public function update_action($id){
-		$this->form_validation->set_rules('nome', 'Nome', 'required');
-		$this->form_validation->set_rules('email', 'Email', 'required');
+        $this->_validationRules();
 		if ($this->form_validation->run() == FALSE){
 			$this->alterar($id);
 		}else{
-			$update = array(
-				'nome'=>$this->input->post('nome'),
-				'email'=>$this->input->post('email'),
-			);
+            $arquivo = $this->input->post('arquivo_aux');
+			if($_FILES['arquivo']['name']){
+                $config = $this->_configsUpload();
+                if (!$this->upload->do_upload('arquivo')){
+                    $this->session->set_flashdata('mensagem',$this->upload->display_errors());
+                    redirect(base_url('contatos'));
+                    exit();
+                }else{
+                    unlink($config['upload_path'].$arquivo);
+                    $upload = $this->upload->data();
+                    $arquivo = $upload['file_name'];
+                }
+            }
+            $update = array(
+                'nome'=>$this->input->post('nome'),
+                'email'=>$this->input->post('email'),
+                'arquivo'=>$arquivo,
+            );
 			if($this->Contatos_model->update($id,$update)){
 				$this->session->set_flashdata('mensagem','Contato alterar com sucesso');
 			}else{
@@ -80,5 +109,19 @@ class Contatos extends CI_Controller {
 			redirect(base_url('contatos'));
 		}
 	}
-
+    final function _configsUpload(){
+        $config['upload_path']          = './uploads/';
+        @mkdir($config['upload_path']);
+        $config['allowed_types']        = 'gif|jpg|png|pdf';
+        $config['max_size']             = 2048;
+        $config['max_width']            = 2048;
+        $config['max_height']           = 2048;
+        $config['encrypt_name']         = true;
+        $this->load->library('upload', $config);
+        return $config;
+    }
+    final function _validationRules(){
+        $this->form_validation->set_rules('nome', 'Nome', 'required');
+        $this->form_validation->set_rules('email', 'Email', 'required|valid_email');
+    }
 }
